230915 제품 수정 페이지 디자인 수정 및 코드 수정 / 기타 불필요 코드 삭제 및 수정

+++ src/main/resources/static/js/admin/usermanage.js
@@ -14,37 +14,41 @@ function get_all_users(){
 }
 
 function create_user_list(userObj) {
-    console.log(userObj)
     tbody.innerHTML = '';
     for (user of userObj) {
-        tbody.insertAdjacentHTML("beforeend", `
-            <tr>
-                <td>${user.email}</td>
-                <td>${user.name}</td>
-                <td>${user.birth}</td>
-                <td>${user.phone}</td>
-                <td>${user.role}</td>
-                <td>
-                    <button onclick="delete_user_data(this)">삭제</button>
-                </td>
-            </tr>
-        `)
+        if(user.email === 'admin') {
+            continue;
+        } else {
+            tbody.insertAdjacentHTML("beforeend", `
+                <tr>
+                    <td>${user.email}</td>
+                    <td>${user.name}</td>
+                    <td>${user.birth}</td>
+                    <td>${user.phone}</td>
+                    <td>${user.role}</td>
+                    <td>
+                        <button onclick="delete_user_data(this)">삭제</button>
+                    </td>
+                </tr>
+            `)
+        }
     }
 }
 
 function delete_user_data(e) {
     let userEmail = e.parentElement.parentElement.firstElementChild.innerText;
-    alert("삭제하시겠습니까?")
-    fetch("/admin/user?userEmail=" + userEmail, {
-        method: "DELETE",
-        headers: {
-            'Content-Type': 'application/json',
-            "X-CSRF-TOKEN": csrfToken
-        },
-        body: JSON.stringify(userEmail)})
-        .then(value => {
-           console.log(value);
-           get_all_users();
-        })
-        .catch(reason => console.log(reason))
+    if (confirm("삭제하시겠습니까?")){
+        fetch("/admin/user?userEmail=" + userEmail, {
+            method: "DELETE",
+            headers: {
+                'Content-Type': 'application/json',
+                "X-CSRF-TOKEN": csrfToken
+            },
+            body: JSON.stringify(userEmail)})
+            .then(value => {
+               console.log(value);
+               get_all_users();
+            })
+            .catch(reason => console.log(reason))
+    }
 }
